alphabetize and abstract
